refactor(logo): move renderAppLogo to lib/app/utils.tsx
Extracts renderAppLogo out of SidebarWrapper so it lives alongside the
other app-level logo utilities.
