+++ web/src/sections/sidebar/SidebarWrapper.tsx
@@ -1,21 +1,10 @@
 "use client";
 
-import React from "react";
 import { SidebarLayouts } from "@opal/layouts";
 import { useShowLogoWhenFolded } from "@/lib/sidebar/hooks";
-import { Logo } from "@/lib/app/components";
+import { renderAppLogo } from "@/lib/app/utils";
 
-/**
- * Renders the app-branded logo for use as the `logo` prop on sidebar primitives.
- * Exported so other sidebar entry points (e.g. AdminSidebar) can reuse it.
- */
-export function renderAppLogo(folded: boolean | undefined): React.ReactNode {
-  return (
-    <div className="px-1">
-      <Logo folded={folded} size={28} />
-    </div>
-  );
-}
+export { renderAppLogo };
 
 export interface SidebarWrapperProps {
   foldable?: boolean;
